(1) Treat floating point consistently as double throughout. (2) Initial rete match/unmatch test passes now.
--HG--
extra : convert_revision : svn%3A849a83f0-50c2-4045-bdd1-8b1f698b4b18/trunk%4013

+++ jsoar/src/org/jsoar/kernel/parser/Lexer.java
@@ -676,7 +676,7 @@ public class Lexer implements LexemeTypes
             try
             {
                 lexeme.type = FLOAT_CONSTANT_LEXEME;
-                lexeme.float_val = Float.valueOf(lexeme.string);
+                lexeme.float_val = Double.valueOf(lexeme.string);
             }
             catch (NumberFormatException e)
             {
